Ginsu::Buffer
Very simple and cheap "buffer" class. It just defines a max amount of data to read at a given time per thread because you can't count on the system to have enough free memory for the user's desired ginsu slice size. The idea is to use this value to read up to that point for a given slice, flush, purge the memory value, and repeat in append-only mode for the file until the slice is done.